refactor: fix folder structure.
